Create Collision Script
This file handles all the user input and updates to the collision calculations. The user can use the arrow key 
and select other keys (see README) to adjust the size of the two spheres, their distance apart, and the angle
of motion of the moving sphere. The direction of movement of the static sphere is indicated by the movementCylinder,
the point of collision is indicated by the collisionPoint, and the location of the moving sphere upon collision
is indicated by the collisionSphere. The distance between the two sphere and the distance the moving sphere
must travel before collision are indicate by Text objects placed in the upper left corner. Program does not currently
provide functionality for the user to adjust the default viewpoint